Add `main.go` inside cmd/declutter folder. And move the `declutter.go` into a seperate pkg dir. The idea is that you want independent binary for your cmd/** files. Having a separate main.go for each cmd/** file will create a different binary.
